refactor: TokenError 경로 수정 (#168)
