+++ src/main/java/com/sudo/railo/auth/exception/TokenError.java
@@ -1,4 +1,4 @@
-package com.sudo.railo.auth.security;
+package com.sudo.railo.auth.exception;
 
 import org.springframework.http.HttpStatus;
