+++ src/main/java/com/sudo/railo/global/exception/GlobalExceptionHandler.java
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 
-import com.sudo.railo.auth.security.TokenError;
+import com.sudo.railo.auth.exception.TokenError;
 import com.sudo.railo.global.exception.error.BusinessException;
 import com.sudo.railo.global.exception.error.ErrorResponse;
 import com.sudo.railo.global.exception.error.GlobalError;
